Moved the implementation of bitstream to a source file.
